moved all new eventApi function from gameModeApi.ts to a new eventApi.ts file
